Refine shared packer state integration
