+++ crates/gui-core/Cargo.toml
@@ -7,3 +7,6 @@ edition = "2021"
 egui = "0.31.1"
 psu-packer = { path = "../psu-packer" }
 serde_json = "1"
+chrono = { version = "0.4", default-features = false, features = ["alloc"] }
+ps2-filetypes = { path = "../ps2-filetypes" }
+tempfile = "3"
