+++ crates/psu-packer-gui/src/lib.rs
@@ -1,4 +1,3 @@
-mod pack_job;
 pub(crate) mod sas_timestamps;
 pub mod state;
 pub mod ui;
