+++ crates/psu-packer-gui/src/sas_timestamps.rs
@@ -1,3 +1,1 @@
-pub(crate) use psu_packer::sas::{
-    planned_timestamp_for_folder, planned_timestamp_for_name, TimestampRules,
-};
+// Legacy module retained for compatibility; timestamp planning logic now lives in gui-core.
